backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
pymysql==1.1.*
alembic==1.15.*
pydantic-settings==2.8.*
python-dotenv==1.1.*
pyyaml==6.0.*
openpyxl==3.1.*
python-multipart==0.0.*
httpx==0.28.*
pytest==8.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.*
